- moved all TSX files (except index.tsx) into components folder - deleted unused CreateEditForm.tsx - moved CountField, LabeledCheckbox & TimeField into its corresponding sub-folders
