+++ src/components/App.tsx
@@ -1,17 +1,17 @@
 import React, { useEffect, useState } from 'react';
 import './App.css';
 import { Route, Switch } from 'react-router-dom';
-import routes from './routes/Routes';
-import AuthRoute from './routes/AuthRoute';
+import routes from '../routes/Routes';
+import AuthRoute from '../routes/AuthRoute';
 import { Box } from '@mui/system';
-import PersistentDrawer from './components/organisms/persistant-drawer/PersistentDrawer';
-import { MenuListItem } from './components/organisms/persistant-drawer/PersistentDrawer.types';
-import { useAuth } from './contexts/AuthContextProvider';
+import PersistentDrawer from './organisms/persistant-drawer/PersistentDrawer';
+import { MenuListItem } from './organisms/persistant-drawer/PersistentDrawer.types';
+import { useAuth } from '../contexts/AuthContextProvider';
 import { useTranslation } from 'react-i18next';
 import { Theme, useMediaQuery } from '@mui/material';
-import useEntityManager from './hooks/useEntityManager';
+import useEntityManager from '../hooks/useEntityManager';
 import { UserProfile } from './pages/account/Account';
-import useFileManager from './hooks/useFileManager';
+import useFileManager from '../hooks/useFileManager';
 
 function App(): JSX.Element {
     const { logout, user } = useAuth();
