+++ src/components/pages/legal/PrivacyPolicyPage.tsx
