+++ src/components/pages/workouts/manage/ManageWorkouts.tsx
@@ -1,21 +1,21 @@
 import { FC, useEffect, useState } from 'react';
 import { Stack, Theme, useMediaQuery } from '@mui/material';
 import { useHistory } from 'react-router-dom';
-import DeleteConfirmationDialog from '../../../components/molecules/delete-confirmation-dialog/DeleteConfirmationDialog';
-import PageTitleActionButton from '../../../components/molecules/page-title-action/PageTitleAction';
-import { Workout } from '../../../model/Workout.model';
-import { getPageIdPrefix } from '../../../util/id-util';
-import * as navigate from '../../../util/navigation-util';
-import Accordion from '../../../components/molecules/accordion/Accordion';
-import { AccordionProp } from '../../../components/molecules/accordion/Accordion.types';
-import ExercisesTimeRepsIcons from '../../../components/organisms/exercises-time-reps-icons/ExercisesTimeRepsIcons';
-import ActionsMenu from '../../../components/molecules/actions-menu/ActionsMenu';
-import ExercisesContent from '../../../components/organisms/exercises-content/ExercisesContent';
-import useEntityManager from '../../../hooks/useEntityManager';
+import DeleteConfirmationDialog from '../../../molecules/delete-confirmation-dialog/DeleteConfirmationDialog';
+import PageTitleActionButton from '../../../molecules/page-title-action/PageTitleAction';
+import { Workout } from '../../../../model/Workout.model';
+import { getPageIdPrefix } from '../../../../util/id-util';
+import * as navigate from '../../../../util/navigation-util';
+import Accordion from '../../../molecules/accordion/Accordion';
+import { AccordionProp } from '../../../molecules/accordion/Accordion.types';
+import ExercisesTimeRepsIcons from '../../../organisms/exercises-time-reps-icons/ExercisesTimeRepsIcons';
+import ActionsMenu from '../../../molecules/actions-menu/ActionsMenu';
+import ExercisesContent from '../../../organisms/exercises-content/ExercisesContent';
+import useEntityManager from '../../../../hooks/useEntityManager';
 import { v4 as uuidv4 } from 'uuid';
-import AddButton from '../../../components/atoms/add-button/AddButton';
-import ResponsiveContainer from '../../../components/organisms/responsive-container/ResponsiveContainer';
-import theme from '../../../theme/theme';
+import AddButton from '../../../atoms/add-button/AddButton';
+import ResponsiveContainer from '../../../organisms/responsive-container/ResponsiveContainer';
+import theme from '../../../../theme/theme';
 import FunctionsIcon from '@mui/icons-material/Functions';
 import { Delete, Edit, PlayArrow } from '@mui/icons-material';
